+++ packages/components-js/projects/partials/scripts/generateInitialStylesPartial.ts
@@ -1,5 +1,15 @@
 import { joinArrayElementsToString, withoutTagsOption } from './utils';
-import { INTERNAL_TAG_NAMES, TAG_NAMES } from '@porsche-design-system/shared';
+import { INTERNAL_TAG_NAMES, TAG_NAMES, getMinifiedCss } from '@porsche-design-system/shared';
+import { Styles } from 'jss';
+import {
+  themeDark,
+  fontWeight,
+  themeLight,
+  fontBehavior,
+  fontFamily,
+  fontHyphenation,
+  fontLineHeight,
+} from '@porsche-design-system/utilities-v2';
 
 const tagNames = joinArrayElementsToString(TAG_NAMES.filter((x) => !INTERNAL_TAG_NAMES.includes(x)));
 
@@ -8,19 +18,77 @@ export const generateInitialStylesPartial = (): string => {
   prefix?: string;
   ${withoutTagsOption}
   format?: Format;
+  // Standard normalize.css styles won't be applied by using this option
+  applyNormalizeStyles?: boolean;
 };
 type GetInitialStylesOptionsFormatHtml = Omit<GetInitialStylesOptions, 'withoutTags'> & { format: 'html' };
 type GetInitialStylesOptionsFormatJsx = Omit<GetInitialStylesOptions, 'withoutTags'> & { format: 'jsx' };
 type GetInitialStylesOptionsWithoutTags = Omit<GetInitialStylesOptions, 'format'>;`;
 
+  const normalizeStyles: Styles = {
+    '@global': {
+      '*': {
+        fontFamily: fontFamily,
+        lineHeight: fontLineHeight,
+        ...fontBehavior,
+      },
+
+      p: {
+        fontWeight: fontWeight.regular,
+        ...fontHyphenation,
+      },
+
+      'h1, h2, h3, h4, h5, h6': {
+        fontWeight: fontWeight.regular,
+      },
+
+      a: {
+        color: 'inherit',
+        textDecoration: 'underline',
+        '@media(hover:hover)': {
+          transition: 'color var(--p-transition-duration, .24s) ease',
+          '&:hover': {
+            color: themeLight.state.hover,
+          },
+        },
+      },
+
+      '@media(hover:hover)': {
+        '[data-theme="dark"] a:hover': {
+          color: themeDark.state.hover,
+        },
+      },
+
+      'b, strong': {
+        fontWeight: fontWeight.bold,
+      },
+
+      'em, i': {
+        fontStyle: 'normal',
+      },
+
+      'a, button, input, select, textarea': {
+        outline: '1px solid transparent',
+        outlineOffset: '1px',
+        '&:focus': {
+          outlineColor: 'currentColor',
+        },
+        '&:focus:not(:focus-visible)': {
+          outlineColor: 'transparent',
+        },
+      },
+    },
+  };
+
   const initialStylesFunction = `export function getInitialStyles(opts?: GetInitialStylesOptionsFormatHtml): string;
 export function getInitialStyles(opts?: GetInitialStylesOptionsFormatJsx): JSX.Element;
 export function getInitialStyles(opts?: GetInitialStylesOptionsWithoutTags): string;
 export function getInitialStyles(opts?: GetInitialStylesOptions): string | JSX.Element {
-  const { prefix, withoutTags, format }: GetInitialStylesOptions = {
+  const { prefix, withoutTags, format, applyNormalizeStyles }: GetInitialStylesOptions = {
     prefix: '',
     withoutTags: false,
     format: 'html',
+    applyNormalizeStyles: false,
     ...opts,
   };
 
@@ -32,7 +100,11 @@ export function getInitialStyles(opts?: GetInitialStylesOptions): string | JSX.E
   const styleProps = { [\`data-pds-initial-styles\$\{prefix ? '-' + prefix : ''\}\`]: '' };
   const styleAttributes = convertPropsToAttributeString(styleProps);
 
-  const styles = prefixedTagNames.join() + '{visibility:hidden}.hydrated,.ssr{visibility:inherit}';
+  const prefixedTagNamesStyles = prefixedTagNames.join() + '{visibility:hidden}.hydrated,.ssr{visibility:inherit}';
+
+  const normalizeStyles = \`${getMinifiedCss(normalizeStyles)}\`;
+
+  const styles = applyNormalizeStyles ? prefixedTagNamesStyles.concat(normalizeStyles) : prefixedTagNamesStyles;
 
   const markup = format === 'html'
     ? \`<style \$\{styleAttributes\}>\${styles}</style>\`