+++ packages/components-js/projects/partials/tests/unit/specs/getInitialStyles.spec.ts
@@ -51,3 +51,9 @@ describe('withoutTags: true', () => {
     expect(result).toMatch(regex);
   });
 });
+
+describe('applyNormalizeStyles: true', () => {
+  it('should return core styles and normalize styles', () => {
+    expect(getInitialStyles({ applyNormalizeStyles: true })).toMatchSnapshot();
+  });
+});