+++ packages/components-react/projects/react-wrapper/src/jsdom-polyfill.ts
@@ -1,0 +1,1 @@
+export * from '@porsche-design-system/components-js/jsdom-polyfill';