+++ packages/components-react/tests/e2e/specs/components.e2e.ts
@@ -29,7 +29,7 @@ it('should stringify object props correctly', async () => {
 
 it('should initialize component deterministically', async () => {
   await goto(page, 'core-initializer');
-  await new Promise((resolve) => setTimeout(resolve, 1000));
+  await page.waitForFunction(() => document.querySelectorAll('p-text-field-wrapper').length === 2);
 
   const [component1, component2] = await page.$$('p-text-field-wrapper');