+++ packages/components/src/components/link-pure/link-pure-styles.spec.ts
@@ -18,7 +18,7 @@ describe('getComponentCss()', () => {
     l: 'x-large',
     xl: 'inherit',
   };
-  const breakpointCustomizableAlignLabel: AlignLabel = {
+  const breakpointCustomizableAlignLabel: BreakpointCustomizable<AlignLabel> = {
     base: 'left',
     xs: 'right',
     s: 'left',
@@ -34,7 +34,7 @@ describe('getComponentCss()', () => {
     size: BreakpointCustomizable<TextSize>;
     weight: TextWeight;
     hideLabel: BreakpointCustomizable<boolean>;
-    alignLabel: AlignLabel;
+    alignLabel: BreakpointCustomizable<AlignLabel>;
     hasSubline: boolean;
     hasSlottedAnchor: boolean;
     theme: ThemeExtendedElectricDark;